requirements.txt
fastapi==0.118.3
deprecated==1.2.18
requests==2.32.5
pydantic==2.12.0
openai==2.3.0
python-dotenv==1.1.1
monocle-apptrace==0.6.5
uvicorn[standard]==0.37.0
# opentelemetry-api==1.37.0
# opentelemetry-sdk==1.37.0
# opentelemetry-instrumentation-fastapi==0.58b0
# opentelemetry-instrumentation-requests==0.58b0
# opentelemetry-instrumentation-openai==0.47.3
# opentelemetry-exporter-jaeger==1.21.0
# opentelemetry-exporter-otlp==1.15.0
